pyproject.toml
[tool.poetry]
name = "fastapi-social"
version = "0.1.0"
description = ""
authors = ["Godswill Umukoro <umukoro6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.109.2"}
psycopg2-binary = "^2.9.9"
sqlalchemy = ">=1.4,<2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
